Move out of shapes. Start adding styles. Move it to dot com.
