+++ apps/dotcom/src/components/MultiplayerEditor.tsx
@@ -8,15 +8,20 @@ import { useSync } from '@tldraw/sync'
 import { useCallback } from 'react'
 import {
 	assertExists,
+	atom,
+	DefaultHelperButtons,
 	DefaultKeyboardShortcutsDialog,
 	DefaultKeyboardShortcutsDialogContent,
 	DefaultMainMenu,
+	DefaultQuickActions,
+	DefaultQuickActionsContent,
 	Editor,
 	EditSubmenu,
 	ExportFileContentSubMenu,
 	ExtrasGroup,
 	PeopleMenu,
 	PreferencesGroup,
+	Timer,
 	TLComponents,
 	Tldraw,
 	TldrawUiButton,
@@ -24,6 +29,8 @@ import {
 	TldrawUiButtonLabel,
 	TldrawUiMenuGroup,
 	TldrawUiMenuItem,
+	TLTimerShapeProps,
+	track,
 	useActions,
 	useEditor,
 	useTranslation,
@@ -47,6 +54,65 @@ import { ShareMenu } from './ShareMenu'
 import { SneakyOnDropOverride } from './SneakyOnDropOverride'
 import { StoreErrorScreen } from './StoreErrorScreen'
 import { ThemeUpdater } from './ThemeUpdater/ThemeUpdater'
+
+const showTimer = atom('timer', false)
+
+const HelperButtons = track(function HelperButtons() {
+	const editor = useEditor()
+	const props = editor.getDocumentSettings().meta
+	const store = useCallback(
+		(newProps: TLTimerShapeProps) => {
+			editor.updateDocumentSettings({
+				meta: { ...props, ...newProps },
+			})
+		},
+		[editor, props]
+	)
+	if (!props || !props.initialTime) return
+
+	return (
+		<>
+			{showTimer.get() && (
+				<div
+					style={{
+						margin: '5px',
+					}}
+				>
+					<Timer props={props as any} editor={editor} store={store} />
+				</div>
+			)}
+			<DefaultHelperButtons />
+		</>
+	)
+})
+
+const QuickActions = track(function QuickActions() {
+	return (
+		<DefaultQuickActions>
+			<DefaultQuickActionsContent />
+			<div
+				style={{
+					backgroundColor: showTimer.get() ? '#ddd' : '',
+				}}
+			>
+				<div
+					style={{
+						margin: '0 5px',
+						height: '100%',
+						display: 'flex',
+						alignItems: 'center',
+						pointerEvents: 'all',
+					}}
+					onClick={() => {
+						showTimer.set(!showTimer.get())
+					}}
+				>
+					<div>🕝</div>
+				</div>
+			</div>
+		</DefaultQuickActions>
+	)
+})
 
 const components: TLComponents = {
 	ErrorFallback: ({ error }) => {
@@ -111,6 +177,8 @@ const components: TLComponents = {
 			</div>
 		)
 	},
+	HelperButtons,
+	QuickActions,
 }
 
 export function MultiplayerEditor({
@@ -124,6 +192,7 @@ export function MultiplayerEditor({
 	useLegacyUrlParams()
 
 	const handleUiEvent = useHandleUiEvents()
+	console.log('roomOpenMode', roomOpenMode)
 	const offset = useGetServerOffset()
 	;(window as any).serverOffset = offset
 
@@ -141,6 +210,19 @@ export function MultiplayerEditor({
 
 	const handleMount = useCallback(
 		(editor: Editor) => {
+			let meta = editor.getDocumentSettings().meta
+			if (!meta.initialTime) {
+				meta = {
+					...meta,
+					initialTime: 30 * 1000,
+					remainingTime: 30 * 1000,
+					state: { state: 'stopped' },
+				}
+				editor.updateDocumentSettings({
+					meta,
+				})
+			}
+
 			if (!isReadonly) {
 				;(window as any).app = editor
 				;(window as any).editor = editor
