+++ packages/tldraw/api-report.md
@@ -109,7 +109,6 @@ import { TLStateNodeConstructor } from '@tldraw/editor';
 import { TLStore } from '@tldraw/editor';
 import { TLStoreSnapshot } from '@tldraw/editor';
 import { TLTextShape } from '@tldraw/editor';
-import { TLTimerShape } from '@tldraw/editor';
 import { TLTimerShapeProps } from '@tldraw/editor';
 import { TLVideoAsset } from '@tldraw/editor';
 import { TLVideoShape } from '@tldraw/editor';
@@ -698,10 +697,10 @@ export const DefaultQuickActions: NamedExoticComponent<TLUiQuickActionsProps>;
 export function DefaultQuickActionsContent(): JSX_2.Element | undefined;
 
 // @public (undocumented)
-export const defaultShapeTools: readonly [typeof TextShapeTool, typeof DrawShapeTool, typeof GeoShapeTool, typeof NoteShapeTool, typeof LineShapeTool, typeof FrameShapeTool, typeof ArrowShapeTool, typeof HighlightShapeTool, typeof TimerShapeTool];
+export const defaultShapeTools: readonly [typeof TextShapeTool, typeof DrawShapeTool, typeof GeoShapeTool, typeof NoteShapeTool, typeof LineShapeTool, typeof FrameShapeTool, typeof ArrowShapeTool, typeof HighlightShapeTool];
 
 // @public (undocumented)
-export const defaultShapeUtils: readonly [typeof TextShapeUtil, typeof BookmarkShapeUtil, typeof DrawShapeUtil, typeof GeoShapeUtil, typeof NoteShapeUtil, typeof LineShapeUtil, typeof FrameShapeUtil, typeof ArrowShapeUtil, typeof HighlightShapeUtil, typeof EmbedShapeUtil, typeof ImageShapeUtil, typeof VideoShapeUtil, typeof TimerShapeUtil];
+export const defaultShapeUtils: readonly [typeof TextShapeUtil, typeof BookmarkShapeUtil, typeof DrawShapeUtil, typeof GeoShapeUtil, typeof NoteShapeUtil, typeof LineShapeUtil, typeof FrameShapeUtil, typeof ArrowShapeUtil, typeof HighlightShapeUtil, typeof EmbedShapeUtil, typeof ImageShapeUtil, typeof VideoShapeUtil];
 
 // @public (undocumented)
 export function DefaultSharePanel(): JSX_2.Element;
@@ -1927,44 +1926,6 @@ export interface TimerProps {
     props: TLTimerShapeProps;
     // (undocumented)
     store(timer: Partial<TLTimerShapeProps>): void;
-}
-
-// @public (undocumented)
-export class TimerShapeTool extends StateNode {
-    // (undocumented)
-    static children(): TLStateNodeConstructor[];
-    // (undocumented)
-    static id: string;
-    // (undocumented)
-    static initial: string;
-    // (undocumented)
-    static isLockable: boolean;
-    // (undocumented)
-    shapeType: string;
-}
-
-// @public (undocumented)
-export class TimerShapeUtil extends ShapeUtil<TLTimerShape> {
-    // (undocumented)
-    canResize(_shape: TLTimerShape): boolean;
-    // (undocumented)
-    component(shape: TLTimerShape): JSX_2.Element;
-    // (undocumented)
-    getDefaultProps(): TLTimerShape['props'];
-    // (undocumented)
-    getGeometry(_shape: TLTimerShape): Geometry2d;
-    // (undocumented)
-    hideResizeHandles(_shape: TLTimerShape): boolean;
-    // (undocumented)
-    hideRotateHandle(_shape: TLTimerShape): boolean;
-    // (undocumented)
-    indicator(): null;
-    // (undocumented)
-    static migrations: TLPropsMigrations;
-    // (undocumented)
-    static props: RecordProps<TLTimerShape>;
-    // (undocumented)
-    static type: "timer";
 }
 
 // @public (undocumented)
