--- packages/tldraw/src/lib/shapes/timer/toolStates/Idle.ts
@@ -1,17 +1,0 @@
-import { StateNode, TLPointerEventInfo } from '@tldraw/editor'
-
-export class Idle extends StateNode {
-	static override id = 'idle'
-
-	override onPointerDown(info: TLPointerEventInfo) {
-		this.parent.transition('pointing', info)
-	}
-
-	override onEnter() {
-		this.editor.setCursor({ type: 'cross', rotation: 0 })
-	}
-
-	override onCancel() {
-		this.editor.setCurrentTool('select')
-	}
-}
